requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.4
six==1.17.0
urllib3==2.2.3
